+++ app/src/test/java/twitterdev/airquality/AirQualityApplicationTest.java
@@ -1,0 +1,12 @@
+package twitterdev.airquality;
+
+import org.junit.Test;
+
+import static org.junit.Assert.assertFalse;
+
+public class AirQualityApplicationTest {
+    @Test
+    public void checkedInBuildLeavesTwitterKitUnconfigured() {
+        assertFalse(AirQualityApplication.isTwitterKitConfigured());
+    }
+}